requirements.txt
Django
djangorestframework
django-import_export
django-filter
djangorestframework-word-filter
django-cors-headers